requirements.txt
Django~=4.2.1
djangorestframework~=3.14.0
drf-spectacular~=0.26.2
python-dotenv~=1.0.0
psycopg2
